+++ crates/theme_fluent/src/prelude.rs
@@ -1,0 +1,1 @@
+pub use crate::{vector_graphics::*, *};