+++ crates/theme_fluent/src/vector_graphics/mod.rs
@@ -1,0 +1,6 @@
+pub use self::segmdl2_assets_font::*;
+
+/// Icons in the `Segoe MDL2 Assets font` are accessible via
+/// their unique codepoints, representing the `glyph`. The codepoints
+/// are mapped to this given constants.
+pub mod segmdl2_assets_font;