+++ crates/widgets/src/tab_widget.rs
@@ -111,7 +111,10 @@ widget!(
         spacing: f64,
 
         /// Sets or shares the close button visibility.
-        close_button: Visibility
+        close_button: Visibility,
+
+        /// Indicates if the widget is hovered by the mouse cursor.
+        hover: bool
     }
 );
 
@@ -339,7 +342,7 @@ impl TabWidgetState {
 
             //Hide current body
             ctx.get_widget(current_tab.1)
-                .set("visibility", Visibility::Hidden);
+                .set("visibility", Visibility::Collapsed);
 
             ctx.get_widget(new_tab.0).set("selected", true);
             toggle_flag("selected", &mut ctx.get_widget(new_tab.0));
@@ -358,8 +361,9 @@ impl TabWidgetState {
         //Create the new tab
         let header = self.create_tab_header(ctx, header_text, body);
 
-        //Set tab body hidden
-        ctx.get_widget(body).set("visibility", Visibility::Hidden);
+        //Set tab body collapsed
+        ctx.get_widget(body)
+            .set("visibility", Visibility::Collapsed);
 
         //Push button to the header container
         ctx.append_child_entity_to(header, self.header_container);
@@ -539,7 +543,7 @@ impl Template for TabWidget {
     fn template(self, id: Entity, ctx: &mut BuildContext) -> Self {
         self.name("TabWidget").style("tab_widget").child(
             Grid::new()
-                .rows(Rows::create().push(32).push("*"))
+                .rows(Rows::create().push(34).push("*"))
                 .child(
                     Stack::new()
                         .id(HEADER_CONTAINER)