+++ crates/widgets/src/text_box.rs
@@ -121,14 +121,14 @@ impl Template for TextBox {
                     .child(
                         Grid::new()
                             .clip(true)
-                            .child(text_block)
                             .child(cursor)
+                            .child(text_block)
                             .build(ctx),
                     )
                     .build(ctx),
             )
             .on_changed("text", move |ctx, _| {
-                ctx.send_message(TextAction::ForceUpdate, text_behavior);
+                ctx.send_message(TextAction::ForceUpdate(false), text_behavior);
             })
     }
 }