+++ crates/widgets/src/toggle_button.rs
@@ -54,7 +54,10 @@ widget!(
         selected: bool,
 
         /// Sets or shares the spacing between icon and text.
-        spacing: f64
+        spacing: f64,
+
+        /// Indicates if the widget is hovered by the mouse cursor.
+        hover: bool
     }
 );